// libs/ort-alumni-client-lib/features/feature-expansion-panel/src/lib/state/panel-action-type.enum.ts
export enum PanelActionType {
  Init = 'init',
  ExcludeStep = 'excludeStep',
  setCurrentStep = 'setCurrentStep',
  nextStep = 'nextStep',
  prevStep = 'prevStep',
  setDisabledStep = 'setDisabledStep',
  rollBackDisabled = 'rollBackDisabled',
  disabledAll = 'disabledAll',
  waitingMode = 'waitingMode'
}